BCH tests remove duplicate code
