+++ test/integration/tx/bitcoin.test.ts
@@ -48,7 +48,7 @@ function testBatchTransaction(chain: Chain) {
   })
 }
 
-function testSweepTransaction(chain: Chain) {
+function testSweepTransaction(chain: Chain, nodeChain: Chain) {
   it('should sweep wallet balance', async () => {
     await fundWallet(chains.bitcoinWithJs)
 
@@ -60,7 +60,7 @@ function testSweepTransaction(chain: Chain) {
 
     let sendTxChain
     if (bal === 0) {
-      sendTxChain = chains.bitcoinWithNode
+      sendTxChain = nodeChain
     } else {
       sendTxChain = chain
     }
@@ -244,7 +244,7 @@ function testSignPSBTScript(chain: Chain) {
   })
 }
 
-function testSignBatchP2SHTransaction(chain: Chain) {
+function testSignBatchP2SHTransaction(chain: Chain, legacy: boolean) {
   it("Should redeem two P2SH's", async () => {
     const network = chain.network
     const value = config[chain.name as keyof typeof config].value
@@ -280,11 +280,13 @@ function testSignBatchP2SHTransaction(chain: Chain) {
       OPS.OP_CHECKMULTISIG
     ])
 
-    const paymentVariantOne = bitcoinJs.payments.p2wsh({
+    const paymentVariant = legacy ? bitcoinJs.payments.p2sh : bitcoinJs.payments.p2wsh
+
+    const paymentVariantOne = paymentVariant({
       redeem: { output: multisigOutputOne, network: network as BitcoinNetwork },
       network: network as BitcoinNetwork
     })
-    const paymentVariantTwo = bitcoinJs.payments.p2wsh({
+    const paymentVariantTwo = paymentVariant({
       redeem: { output: multisigOutputTwo, network: network as BitcoinNetwork },
       network: network as BitcoinNetwork
     })
@@ -317,7 +319,8 @@ function testSignBatchP2SHTransaction(chain: Chain) {
 
     txb.addInput(initiationTx.hash, multiOne.multiVout.n, 0, paymentVariantOne.output)
     txb.addInput(initiationTx.hash, multiTwo.multiVout.n, 0, paymentVariantTwo.output)
-    txb.addOutput(unusedAddressTwo, value.toNumber() * 2 - txfee)
+    const addr = unusedAddressTwo.toString()
+    txb.addOutput(BitcoinUtils.addrToBitcoinJS(addr, network as BitcoinNetwork), value.toNumber() * 2 - txfee)
 
     const tx = txb.buildIncomplete()
 
@@ -339,7 +342,7 @@ function testSignBatchP2SHTransaction(chain: Chain) {
       [addresses[0].address, addresses[0].address],
       tx,
       0,
-      true
+      !legacy
     )
 
     const signaturesTwo = await chain.client.getMethod('signBatchP2SHTransaction')(
@@ -360,7 +363,7 @@ function testSignBatchP2SHTransaction(chain: Chain) {
       [addresses[1].address, addresses[1].address],
       tx,
       0,
-      true
+      !legacy
     )
 
     const multiOneInput = bitcoinJs.script.compile([OPS.OP_0, signaturesOne[0], signaturesTwo[0]])
@@ -370,11 +373,19 @@ function testSignBatchP2SHTransaction(chain: Chain) {
     multiOne.paymentParams = { redeem: { output: multisigOutputOne, input: multiOneInput, network }, network }
     multiTwo.paymentParams = { redeem: { output: multisigOutputTwo, input: multiTwoInput, network }, network }
 
-    multiOne.paymentWithInput = bitcoinJs.payments.p2wsh(multiOne.paymentParams)
-    multiTwo.paymentWithInput = bitcoinJs.payments.p2wsh(multiTwo.paymentParams)
+    if (legacy) {
+      multiOne.paymentWithInput = bitcoinJs.payments.p2sh(multiOne.paymentParams)
+      multiTwo.paymentWithInput = bitcoinJs.payments.p2sh(multiTwo.paymentParams)
 
-    tx.setWitness(0, multiOne.paymentWithInput.witness)
-    tx.setWitness(1, multiTwo.paymentWithInput.witness)
+      tx.setInputScript(0, multiOne.paymentWithInput.input)
+      tx.setInputScript(1, multiTwo.paymentWithInput.input)
+    } else {
+      multiOne.paymentWithInput = bitcoinJs.payments.p2wsh(multiOne.paymentParams)
+      multiTwo.paymentWithInput = bitcoinJs.payments.p2wsh(multiTwo.paymentParams)
+
+      tx.setWitness(0, multiOne.paymentWithInput.witness)
+      tx.setWitness(1, multiTwo.paymentWithInput.witness)
+    }
 
     const claimTxHash = await chain.client.getMethod('sendRawTransaction')(tx.toHex())
 
@@ -403,7 +414,7 @@ describe('Transactions', function () {
     testBatchTransaction(chains.bitcoinWithLedger)
     testSignPSBTSimple(chains.bitcoinWithLedger)
     testSignPSBTScript(chains.bitcoinWithLedger)
-    testSignBatchP2SHTransaction(chains.bitcoinWithLedger)
+    testSignBatchP2SHTransaction(chains.bitcoinWithLedger, false)
   })
 
   describe('Bitcoin - Node', () => {
@@ -411,7 +422,8 @@ describe('Transactions', function () {
     testBatchTransaction(chains.bitcoinWithNode)
     testSignPSBTSimple(chains.bitcoinWithNode)
     testSignPSBTScript(chains.bitcoinWithNode)
-    testSignBatchP2SHTransaction(chains.bitcoinWithNode)
+    testSignBatchP2SHTransaction(chains.bitcoinWithNode, false)
+    testSignBatchP2SHTransaction(chains.bitcoinWithNode, true)
   })
 
   describe('Bitcoin - Js', () => {
@@ -423,7 +435,25 @@ describe('Transactions', function () {
     testBatchTransaction(chains.bitcoinWithJs)
     testSignPSBTSimple(chains.bitcoinWithJs)
     testSignPSBTScript(chains.bitcoinWithJs)
-    testSignBatchP2SHTransaction(chains.bitcoinWithJs)
-    testSweepTransaction(chains.bitcoinWithJs)
+    testSignBatchP2SHTransaction(chains.bitcoinWithJs, false)
+    testSignBatchP2SHTransaction(chains.bitcoinWithJs, true)
+    testSweepTransaction(chains.bitcoinWithJs, chains.bitcoinWithNode)
+  })
+
+  describe('Bitcoin Cash - Node', () => {
+    testTransaction(chains.bitcoinCashWithNode)
+    testBatchTransaction(chains.bitcoinCashWithNode)
+    testSignBatchP2SHTransaction(chains.bitcoinCashWithNode, true)
+  })
+
+  describe('Bitcoin Cash - Js', () => {
+    before(async function () {
+      await importBitcoinAddresses(chains.bitcoinCashWithJs)
+      await fundWallet(chains.bitcoinCashWithJs)
+    })
+    testTransaction(chains.bitcoinCashWithJs)
+    testBatchTransaction(chains.bitcoinCashWithJs)
+    testSignBatchP2SHTransaction(chains.bitcoinCashWithJs, true)
+    testSweepTransaction(chains.bitcoinCashWithJs, chains.bitcoinCashWithNode)
   })
 })
